replace prepare with prepublishOnly
The old `prepare` was running before a publish but also when we `npm
install`. This is not so good and breaks builds for windows users
(#121). A fix for this is to only run the code before an actual publish.
This can be configured with the `prepublishOnly` script name.

See: https://docs.npmjs.com/misc/scripts#deprecation-note

Fixes: #121