ci: add provenance to published packages (#1145)
This commit adds provenance for all published packages. See the NPM documentation [0].

Provenance will allow people to verify that the packages were actually built on GH Actions and with the content of the corresponding commit. This will help with supply chain security.

For this to work, the `id-token` permission was added only where necessary.
Additionally, the repository link was added to the `package.json` for this to work.

[0]: https://docs.npmjs.com/generating-provenance-statements