Remove clip-path from 20/arrow-left-circle
At size 20, the `arrow-left-circle` icon was not displaying correctly in my Elixir Phoenix LiveView application.

Comparing to the other `arrow-*-circle` icons of size 20, as well as the other sizes of `arrow-left-circle`, this is the only variant that contains a `clip-path`.

Removing the clip-path allows the icon to display correctly.